refactor: 💡 Remove v1 types and queries
