+++ .eslintrc
@@ -101,5 +101,5 @@
     "project": "./tsconfig.json"
   },
   "extends": ["standard", "semistandard", "plugin:@typescript-eslint/recommended", "prettier"],
-  "ignorePatterns": ["src/polkadot/", "src/middleware/types.ts", "src/middleware/typesV2.ts", "docs/*"]
+  "ignorePatterns": ["src/polkadot/", "src/middleware/types.ts", "docs/*"]
 }
