--- codegen.yml
@@ -1,14 +1,0 @@
-schema:
-  - ${MIDDLEWARE_LINK}:
-      headers:
-        x-api-key: ${MIDDLEWARE_KEY}
-generates:
-  ./src/middleware/types.ts:
-    config:
-      defaultMapper: any
-      scalars:
-        DateTime: string
-        JSON: '{ [key: string]: any }'
-        CustomBoolean: boolean
-    plugins:
-      - 'typescript'
