+++ src/types/utils/index.ts
@@ -84,9 +84,7 @@ export type WithRequired<T, K extends keyof T> = T & { [P in K]-?: T[P] };
 /**
  * Pick a single property from T and ensure it is defined
  */
-export type Ensured<T, K extends keyof T> = Required<Pick<T, K>>;
-
-export type EnsuredV2<T, K extends keyof T> = Required<Pick<T, K>> & {
+export type Ensured<T, K extends keyof T> = Required<Pick<T, K>> & {
   [SubKey in K]: NonNullable<T[SubKey]>;
 };
